pyproject.toml
[project]
name = "demo-priority-queue"
version = "0.1.0"
description = "Its a demo project for priority queue for my notifications module"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.50",
    "python-dotenv>=1.0.1",
]
